Move grid into a utils file
